MGMT-14737: Add base terraform controller for both vsphere and nutanix.
