+++ src/assisted_test_infra/test_infra/controllers/node_controllers/__init__.py
@@ -5,4 +5,11 @@ from .node_controller import NodeController
 from .terraform_controller import TerraformController
 from .vsphere_controller import VSphereController
 
-__all__ = ["TerraformController", "NodeController", "VSphereController", "Disk", "Node", "LibvirtController"]
+__all__ = [
+    "TerraformController",
+    "NodeController",
+    "VSphereController",
+    "Disk",
+    "Node",
+    "LibvirtController",
+]
