+++ src/assisted_test_infra/test_infra/controllers/node_controllers/vsphere_controller.py
@@ -1,31 +1,24 @@
 import os
 import time
-from typing import Any, Callable, Dict, List, Tuple
+from typing import Any, Callable, List, Tuple
 
-import libvirt
 from pyVim import task
 from pyVim.connect import Disconnect, SmartConnect
 from pyVmomi import vim
 
-from assisted_test_infra.test_infra import BaseClusterConfig
-from assisted_test_infra.test_infra.controllers.node_controllers.disk import Disk
 from assisted_test_infra.test_infra.controllers.node_controllers.node import Node
-from assisted_test_infra.test_infra.controllers.node_controllers.node_controller import NodeController
-from assisted_test_infra.test_infra.helper_classes.config.base_vsphere_config import BaseVSphereConfig
-from assisted_test_infra.test_infra.tools import terraform_utils
-from assisted_test_infra.test_infra.utils import TerraformControllerUtil, utils
+from assisted_test_infra.test_infra.controllers.node_controllers.tf_controller import TFController
+from assisted_test_infra.test_infra.utils import utils
 from service_client import log
 
 
-class VSphereController(NodeController):
-    def __init__(self, config: BaseVSphereConfig, cluster_config: BaseClusterConfig):
-        super().__init__(config, cluster_config)
-        self.cluster_name = cluster_config.cluster_name.get()
-        folder = TerraformControllerUtil.create_folder(self.cluster_name, platform=config.tf_platform)
-        self.tf = terraform_utils.TerraformUtils(working_dir=folder, terraform_init=False)
+class VSphereController(TFController):
+    def _get_provider_client(self) -> object:
+        return None
 
-    def get_all_vars(self):
-        return {**self._config.get_all(), **self._entity_config.get_all(), "cluster_name": self.cluster_name}
+    @property
+    def terraform_vm_resource_type(self) -> str:
+        return "vsphere_virtual_machine"
 
     def prepare_nodes(self):
         if not os.path.exists(self._entity_config.iso_download_path):
@@ -45,26 +38,14 @@ class VSphereController(NodeController):
         config = self.get_all_vars()
         self.tf.set_and_apply(**config)
 
-    def list_nodes(self) -> List[Node]:
-        vms = self.__get_vms()
-
-        def vsphere_vm_to_node(terraform_vm_state):
-            return Node(
-                name=terraform_vm_state["attributes"]["name"],
-                private_ssh_key_path=self._config.private_ssh_key_path,
-                node_controller=self,
-            )
-
-        return list(map(vsphere_vm_to_node, vms))
-
     def get_cpu_cores(self, node_name: str) -> int:
-        return self.__get_vm(node_name)["attributes"]["num_cpus"]
+        return self._get_vm(node_name)["attributes"]["num_cpus"]
 
     def get_ram_kib(self, node_name: str) -> int:
-        return self.__get_vm(node_name)["attributes"]["memory"] * 1024
+        return self._get_vm(node_name)["attributes"]["memory"] * 1024
 
     def get_node_ips_and_macs(self, node_name) -> Tuple[List[str], List[str]]:
-        vm_attributes = self.__get_vm(node_name)["attributes"]
+        vm_attributes = self._get_vm(node_name)["attributes"]
 
         ips = vm_attributes["guest_ip_addresses"]
         macs = []
@@ -87,7 +68,7 @@ class VSphereController(NodeController):
         nodes = self.list_nodes()
 
         for node in nodes:
-            self.start_node(node.name)
+            self.start_node(node.name, False)
 
         return self.list_nodes()
 
@@ -111,98 +92,10 @@ class VSphereController(NodeController):
 
         self.__run_on_vm(node_name, reboot)
 
-    def get_ingress_and_api_vips(self) -> dict:
-        if self._entity_config.api_vip and self._entity_config.ingress_vip:
-            return {"api_vip": self._entity_config.api_vip, "ingress_vip": self._entity_config.ingress_vip}
-
-        return None
-
-    def set_dns(self, api_ip: str, ingress_ip: str) -> None:
-        utils.add_dns_record(
-            cluster_name=self.cluster_name,
-            base_dns_domain=self._entity_config.base_dns_domain,
-            api_ip=api_ip,
-            ingress_ip=ingress_ip,
-        )
-
-    def format_node_disk(self, node_name: str, disk_index: int = 0) -> None:
-        raise NotImplementedError
-
-    def list_disks(self, node_name: str) -> List[Disk]:
-        raise NotImplementedError
-
     def is_active(self, node_name) -> bool:
         # TODO[vrutkovs]: use vSphere API to determine if node is running
         # Currently its assumed to be always on
         return True
-
-    def format_all_node_disks(self) -> None:
-        raise NotImplementedError
-
-    def list_networks(self) -> List[Any]:
-        raise NotImplementedError
-
-    def list_leases(self, network_name: str) -> List[Any]:
-        raise NotImplementedError
-
-    def attach_test_disk(self, node_name: str, disk_size: int, bootable=False, persistent=False, with_wwn=False):
-        raise NotImplementedError
-
-    def detach_all_test_disks(self, node_name: str):
-        raise NotImplementedError
-
-    def get_cluster_network(self) -> str:
-        raise NotImplementedError
-
-    def setup_time(self) -> str:
-        raise NotImplementedError
-
-    def set_boot_order(self, node_name, cd_first=False) -> None:
-        raise NotImplementedError
-
-    def set_per_device_boot_order(self, node_name, key: Callable[[Disk], int]) -> None:
-        raise NotImplementedError
-
-    def get_host_id(self, node_name: str) -> str:
-        raise NotImplementedError
-
-    def set_cpu_cores(self, node_name: str, core_count: int) -> None:
-        raise NotImplementedError
-
-    def set_ram_kib(self, node_name: str, ram_kib: int) -> None:
-        raise NotImplementedError
-
-    def attach_interface(self, node_name, network_xml: str) -> Tuple[libvirt.virNetwork, str]:
-        raise NotImplementedError
-
-    def add_interface(self, node_name, network_name, target_interface: str) -> str:
-        raise NotImplementedError
-
-    def undefine_interface(self, node_name: str, mac: str):
-        raise NotImplementedError
-
-    def create_network(self, network_xml: str) -> libvirt.virNetwork:
-        raise NotImplementedError
-
-    def get_network_by_name(self, network_name: str) -> libvirt.virNetwork:
-        raise NotImplementedError
-
-    def destroy_network(self, network: libvirt.virNetwork):
-        raise NotImplementedError
-
-    def set_single_node_ip(self, ip):
-        raise NotImplementedError
-
-    def __get_vm(self, node_name: str) -> Dict[str, Any]:
-        return next((vm for vm in self.__get_vms() if vm["attributes"]["name"] == node_name), None)
-
-    def __get_vms(self) -> List[Dict[str, Any]]:
-        vms_object_type = self.tf.get_resources(resource_type="vsphere_virtual_machine")
-
-        if not vms_object_type:
-            return list()
-
-        return [vm_instance for vms_objects in vms_object_type for vm_instance in vms_objects["instances"]]
 
     def _create_folder(self, name: str):
         connection = self.__new_connection()
